requirements.txt
alabaster==0.7.12
alembic==1.0.6
asn1crypto==0.24.0
astroid==2.1.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.6
bleach==3.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
cryptography==2.6.1
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==17.1.1
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
numpydoc==0.8.0
packaging==19.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==5.4.8
psycopg2-binary==2.7.7
pyasn1==0.4.5
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
PyNaCl==1.3.0
pyparsing==2.3.1
PyQt5==5.9.2
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
pywin32-ctypes==0.2.0
pyzmq==17.1.2
QtAwesome==0.5.5
qtconsole==4.4.3
QtPy==1.6.0
requests==2.21.0
rope==0.11.0
sip==4.19.8
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.16
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.2.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wfastcgi==3.0.0
wrapt==1.11.1
